backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
ruff==0.5.7
sqlparse==0.4.4
typing_extensions==4.11.0
uvicorn==0.30.1
whitenoise==6.6.0
